requirements.txt
numpy>=1.24,<2.0
openai>=1.6.0,<2.0
python-dotenv>=1.0.0,<2.0
fastapi>=0.110.0,<1.0
uvicorn>=0.23.0,<1.0
